Refactor the MonkeyBusinessManager
Having a separate class with a dependency wrapper required some
weird machinations.  It's cleaner to have a single global singleton
class that speaks the FastAPI dependency protocol.  As a bonus,
it can now be initialized and cleaned up repeatedly over the course
of the test suite without being recreated.
